Moved the API logic from step defs to the page object file
